Cleanup onshape paths
