+++ src/backend/onshape-api/endpoints/thumbnails.ts
@@ -5,7 +5,7 @@ import {
     InstancePath,
     toElementApiPath,
     toInstanceApiPath
-} from "../../../shared/path";
+} from "../../../shared/onshape-path";
 import { apiPath } from "../api-path";
 
 /** Represents the possible sizes of a thumbnail. */
