+++ src/backend/onshape-api/endpoints/versions.ts
@@ -5,7 +5,7 @@ import {
     InstancePath,
     toDocumentApiPath,
     toInstanceApiObject
-} from "../../../shared/path";
+} from "../../../shared/onshape-path";
 import { apiPath } from "../api-path";
 
 /**
