+++ src/frontend/api-utils/library.ts
@@ -12,6 +12,7 @@ export function toLibraryPath(library: Library): string {
 export function toInsertablePath(insertableId: string): string {
     return `/insertable/${insertableId}`;
 }
+
 export function getLibraryName(library: string): string {
     switch (library) {
         case Library.FRC_DESIGN_LIB:
