+++ src/frontend/insert/insert-hooks.ts
@@ -3,7 +3,7 @@ import { useSearch } from "@tanstack/react-router";
 import { apiPost } from "../api-utils/api";
 import { InsertableOut } from "../../shared/api-models";
 import { ElementType } from "../../shared/types";
-import { toElementApiPath } from "../../shared/path";
+import { toElementApiPath } from "../../shared/onshape-path";
 import { showLoadingToast, showSuccessToast } from "../common/toaster";
 import { queryClient } from "../query-client";
 import { getAppErrorHandler } from "../api-utils/errors";
